fix: prevent division by zero in noise model calculation

+++ ros2/sycl_points_ros2/include/sycl_points_ros2/declare_lidar_odometry_params.hpp
@@ -288,15 +288,16 @@ inline pipeline::lidar_odometry::Parameters declare_lidar_odometry_parameters(rc
         // sigma_range = sigma_grazing = 0 reproduces the standard {1e-3, 1, 1} behavior.
         {
             auto& nm = solver.source_noise_model;
-            nm.sigma_range = node->declare_parameter<double>("registration/source_noise_model/sigma_range",
-                                                             nm.sigma_range);
-            nm.sigma_grazing = node->declare_parameter<double>("registration/source_noise_model/sigma_grazing",
-                                                               nm.sigma_grazing);
-            nm.min_eigenvalue = node->declare_parameter<double>("registration/source_noise_model/min_eigenvalue",
-                                                                nm.min_eigenvalue);
-            nm.max_eigenvalue = node->declare_parameter<double>("registration/source_noise_model/max_eigenvalue",
-                                                                nm.max_eigenvalue);
-            nm.cos_min = node->declare_parameter<double>("registration/source_noise_model/cos_min", nm.cos_min);
+            nm.sigma_range = static_cast<float>(
+                node->declare_parameter<double>("registration/source_noise_model/sigma_range", nm.sigma_range));
+            nm.sigma_grazing = static_cast<float>(
+                node->declare_parameter<double>("registration/source_noise_model/sigma_grazing", nm.sigma_grazing));
+            nm.min_eigenvalue = static_cast<float>(
+                node->declare_parameter<double>("registration/source_noise_model/min_eigenvalue", nm.min_eigenvalue));
+            nm.max_eigenvalue = static_cast<float>(
+                node->declare_parameter<double>("registration/source_noise_model/max_eigenvalue", nm.max_eigenvalue));
+            nm.cos_min = static_cast<float>(
+                node->declare_parameter<double>("registration/source_noise_model/cos_min", nm.cos_min));
         }
         // Rotation Constraint
         {
